Implemented Dekel-Zhao halo profile
Also added to fitting wrapper.
Also added possible tied functions between M*/Mvir and s1/c2 (which could also be implicitly tied to fDM if Mvir is determined from fDM).
Moved calc_rvir() to be a method of DarkMatterHalo only, as this method was the same for all inheriting halo classes.